% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_format_helper.R
\name{age_format_helper}
\alias{age_format_helper}
\title{Helper function for formatting yyy-mm ages as years}
\usage{
age_format_helper(x)
}
\arguments{
\item{x}{A character vector of age(s) in yyy-mm format (e.g. 13-06) to be converted to years (13.5)}
}
\description{
Helper function for formatting yyy-mm ages as years
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
