% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_events.R
\name{clean_events}
\alias{clean_events}
\title{Regularizes event names}
\usage{
clean_events(x)
}
\arguments{
\item{x}{a character vector of event names}
}
\value{
a character vector of event names with naming conventions enforced to
  regularize event names
}
\description{
XXX
}
