% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_many.R
\name{coalesce_many}
\alias{coalesce_many}
\title{Combined paired sets of columns following a join operation}
\usage{
coalesce_many(df)
}
\arguments{
\item{df}{a data frame following a join and thereby containing paired columns
of the form Col_1.x, Col_1.y}
}
\value{
returns a data frame with all sets of paired columns combined into
  single columns and named as, for example, Col_1, Col_2 etc.
}
\description{
Combined paired sets of columns following a join operation
}
\seealso{
\code{coalesce_many} runs inside
  \code{\link{swim_parse_splash}}
}
