% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_relay_swimmers.R
\name{collect_relay_swimmers}
\alias{collect_relay_swimmers}
\title{Collects relay swimmers as a data frame within \code{swim_parse}}
\usage{
collect_relay_swimmers(x, typo_2 = typo, replacement_2 = replacement)
}
\arguments{
\item{x}{output from \code{read_results} followed by \code{add_row_numbers}}

\item{typo_2}{list of typos from \code{swim_parse}}

\item{replacement_2}{list of replacements for typos from \code{swim_parse}}
}
\value{
returns a data frame of relay swimmers and the associated performance row number
}
\description{
Collects relay swimmers as a data frame within \code{swim_parse}
}
\seealso{
\code{collect_relay_swimmers} runs inside of \code{swim_parse}
}
