% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_relay_swimmers_omega.R
\name{collect_relay_swimmers_omega}
\alias{collect_relay_swimmers_omega}
\title{Collects relay swimmers as a data frame within \code{swim_parse_omega}}
\usage{
collect_relay_swimmers_omega(x)
}
\arguments{
\item{x}{output from \code{read_results} followed by \code{add_row_numbers}}
}
\value{
returns a data frame of relay swimmers and the associated performance row number
}
\description{
Collects relay swimmers as a data frame within \code{swim_parse_omega}
}
\seealso{
\code{collect_relay_swimmers_data} runs inside of \code{swim_parse_omega}
}
