% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splash_collect_relay_swimmers.R
\name{collect_relay_swimmers_splash}
\alias{collect_relay_swimmers_splash}
\title{Collects relay swimmers as a data frame within \code{swim_parse_splash}}
\usage{
collect_relay_swimmers_splash(x, relay_indent = Indent_Length)
}
\arguments{
\item{x}{output from \code{read_results} followed by \code{add_row_numbers}}

\item{relay_indent}{the number of spaces relay swimmer lines are indented
compared to regular swimmer lines}
}
\value{
returns a data frame of relay swimmers and the associated performance
  row number
}
\description{
Collects relay swimmers as a data frame within \code{swim_parse_splash}
}
\seealso{
\code{collect_relay_swimmers_data} runs inside of
  \code{swim_parse_splash}
}
