% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_splits_helper.R
\name{correct_splits_helper}
\alias{correct_splits_helper}
\title{Changes lengths associated with splits to new values}
\usage{
correct_splits_helper(df_helper, new_split_length_helper)
}
\arguments{
\item{df_helper}{a dataframe having some split columns (Split_50, Split_100 etc.)}

\item{new_split_length_helper}{split length to rename split columns based on}
}
\value{
a dataframe where all values have been pushed left, replacing `NA`s, and all columns containing only `NA`s have been removed
}
\description{
Useful for dealing with meets where some events are split by 50 and others by 25.
}
\seealso{
\code{fill_left} is a helper function inside \code{lines_sort} and \code{splits_parse}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
