% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_entries.R
\name{determine_entries_helper_2}
\alias{determine_entries_helper_2}
\title{Assign overpowered entries}
\usage{
determine_entries_helper_2(
  i,
  df_helper,
  df_2_helper,
  athlete_ranks_helper,
  athletes_remaining_helper,
  events_competed_helper = Events_Competed,
  max_ind_entries = 3
)
}
\arguments{
\item{i}{a sequential list of numbers incremented by 1.  Used to index
function.}

\item{df_helper}{a data frame of times for the team to be entered.  Must
contain column \code{Event} with the same event naming convention as
\code{op_df}, a column with name matching \code{time_col} containing times
or diving scores, and a column called \code{Name} containing athlete names}

\item{max_ind_entries}{a numeric value denoting the maximum number of
individual events that may be entered by a single athlete}

\item{op_df_helper}{a data frame containing the opponent lineup.  Must
contain column \code{Event} with the same event naming convention as
\code{df}, a column with name matching \code{time_col} containing times or
diving scores, and a column called \code{Name} containing athlete names}
}
\value{
xxx
}
\description{
Assign overpowered entries
}
