% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splash_determine_indent_length.R
\name{determine_indent_length_splash}
\alias{determine_indent_length_splash}
\title{Determines indent length for data within \code{swim_parse_splash}}
\usage{
determine_indent_length_splash(x, time_score_string)
}
\arguments{
\item{x}{output from \code{read_results} followed by \code{add_row_numbers}}
}
\value{
returns a number indicating the number of spaces preceding an
  athlete's name in a tie row
}
\description{
In Splash results there are two line types that are of interest and don't
begin with either a place or a special string (DNS, DSQ etc.).  These are
ties and relays swimmers.  Relay swimmers are indented further than ties.
This function determines the number of spaces, called indent length, prior to
a tie row, plus a pad of four spaces.
}
\seealso{
\code{determine_indent_length_splash} runs inside of
  \code{swim_parse_splash}
}
