% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_errors.R
\name{discard_errors}
\alias{discard_errors}
\title{Discards elements of list that have an error value from \code{purrr::safely}}
\usage{
discard_errors(x)
}
\arguments{
\item{x}{a list of lists from \code{purrr::map} and \code{purrr:safely}}
}
\value{
a list of lists where elements with an error have been discarded and all error elements have been removed
}
\description{
Discards elements of list that have an error value from \code{purrr::safely}
}
\references{
https://pilgrim.netlify.app/post/scrapping-websites-and-building-a-large-dataset-with-swimmer/
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
