% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results.R
\name{format_results}
\alias{format_results}
\title{Formats data for analysis within \code{swim_parse}}
\usage{
format_results(df)
}
\arguments{
\item{df}{a dataframe of results at the end of \code{swim_parse}}
}
\value{
returns a formated dataframe
}
\description{
Takes the output of \code{read_results} and, inside of \code{swim_parse}, removes "special" strings like DQ and SCR from results, replacing them with NA.  Also ensures that all athletes have a Finals_Time, by moving over Prelims_Time.  This makes later analysis much easier.
}
\seealso{
\code{splits_parse} runs inside \code{\link{swim_parse}} on the output of \code{\link{read_results}} with row numbers from \code{\link{add_row_numbers}}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
