% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%notin\%}
\alias{\%notin\%}
\title{not_in
Helper function for negating the %in% operator}
\usage{
... \%notin\% NA
}
\description{
not_in
Helper function for negating the %in% operator
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
