% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_parse_omega.R
\name{heat_parse_omega}
\alias{heat_parse_omega}
\title{Pulls out heat labels from text}
\usage{
heat_parse_omega(text)
}
\arguments{
\item{text}{output from \code{read_results} followed by
\code{add_row_numbers}}
}
\value{
returns a data frame with heat names and row numbers to eventually
  be recombined with swimming results inside \code{swim_parse_omega}
}
\description{
Locates heat labels in text of results output from \code{read_results} and
their associated row numbers.  The resulting data frame is joined back into
results to include heat numbers
}
\seealso{
\code{heat_parse_omega} is a helper function inside
  \code{\link{swim_parse_omega}}
}
