% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy3_places.R
\name{hy3_places}
\alias{hy3_places}
\title{Helper for reading prelims and finals places from Hy-Tek .hy3 files}
\usage{
hy3_places(
  file,
  type = c("prelims", "relay_prelims", "finals", "relay_finals")
)
}
\arguments{
\item{file}{an output of read_results, from an .hy3 file}

\item{type}{type of times, either "prelims", "relay_prelims", "finals" or "relay_finals"}
}
\value{
a data frame where column 1 is times and column 2 is row number
}
\description{
Used to pull prelims and finals places from .hy3 files as part of parsing them.
}
\seealso{
\code{hy3_places} is run inside of \code{\link{hy3_parse}}
}
