% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy3_times.R
\name{hy3_times}
\alias{hy3_times}
\title{Helper for reading prelims and finals times from Hy-Tek .hy3 files}
\usage{
hy3_times(file, type = c("prelims", "relay_prelims", "finals", "relay_finals"))
}
\arguments{
\item{file}{an output of read_results, from an .hy3 file}

\item{type}{type of times, either "prelims", "relay_prelims", "finals" or "relay_finals"}
}
\value{
a dataframe where column 1 is times and column 2 is row number
}
\description{
Used to pull prelims and finals times from .hy3 files as part of parsing them.
}
\seealso{
\code{hy3_times} is run inside of \code{\link{hy3_parse}}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
