% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytek_clean_strings.R
\name{hytek_clean_strings}
\alias{hytek_clean_strings}
\title{Cleans input strings}
\usage{
hytek_clean_strings(x, time_score_string = Time_Score_String)
}
\arguments{
\item{x}{a list of character strings}

\item{time_score_string}{a regex string for matching results (times and
scores) but not special strings like DQ}
}
\value{
returns a list of character strings that have been cleaned in
  preparation for parsing/sorting

  #' @seealso \code{hytek_clean_strings} runs inside of
  \code{hytek_parse_splash}
}
\description{
Cleans input from \code{read_results} is passed to \code{hytek_swim_parse}
to remove unnneded characters and otherwise set it up for sorting.  Input is
in the form of character strings
}
