% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytek_list_sorters.R
\name{hytek_length_3_DQ_sort}
\alias{hytek_length_3_DQ_sort}
\title{Sorts data in DQ lists of length 3 within \code{hytek_swim_parse}}
\usage{
hytek_length_3_DQ_sort(x)
}
\arguments{
\item{x}{a list of lists containing DQ results with all sub-lists having
length 3 strings}
}
\value{
returns a formatted data frame to be combined with others to make the
  output of \code{hytek_swim_parse}
}
\description{
XXXXXX
}
