% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytek_list_sorters.R
\name{hytek_length_4_sort}
\alias{hytek_length_4_sort}
\title{Sorts data in lists of length 4 within \code{hytek_swim_parse}}
\usage{
hytek_length_4_sort(x, time_score_specials_string = Time_Score_Specials_String)
}
\arguments{
\item{x}{a list of lists with all sub-lists having length 4
strings}

\item{time_score_specials_string}{a regex string for matching results - i.e.
times, diving scores and 'specials' like DQ}
}
\value{
returns a formatted data frame to be combined with others to make the
  output of \code{hytek_swim_parse}
}
\description{
XXXXXX
}
