% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytek_list_sorters.R
\name{hytek_length_5_sort}
\alias{hytek_length_5_sort}
\title{Sorts data in lists of length 5 within \code{hytek_swim_parse}}
\usage{
hytek_length_5_sort(
  x,
  name_string = Name_String,
  age_string = Age_String,
  para_string = Para_String,
  time_score_specials_string = Time_Score_Specials_String
)
}
\arguments{
\item{x}{a list of lists with all sub-lists having length 5
strings}

\item{name_string}{a regex string for matching athlete names}

\item{age_string}{a regex string for matching athlete ages}

\item{para_string}{a regex string for matching Paralympics classification
strings}

\item{time_score_specials_string}{a regex string for matching results - i.e.
times, diving scores and 'specials' like DQ}
}
\value{
returns a formatted data frame to be combined with others to make the
  output of \code{hytek_swim_parse}
}
\description{
XXXXXX
}
