% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hytek_list_sorters.R
\name{hytek_length_8_sort}
\alias{hytek_length_8_sort}
\title{Sort data in lists of length 8 within \code{hytek_swim_parse}}
\usage{
hytek_length_8_sort(
  x,
  brit_id_string = Brit_ID_String,
  para_string = Para_String,
  age_string = Age_String,
  time_score_specials_string = Time_Score_Specials_String
)
}
\arguments{
\item{x}{a list of lists with all sub-lists having length 8}

\item{brit_id_string}{a regex string for matching British swimming IDs}

\item{para_string}{a regex string for matching Paralympics classification
strings}

\item{age_string}{a regex string for matching athlete ages}

\item{time_score_specials_string}{a regex string for matching results - i.e.
times, diving scores and 'specials' like DQ}
}
\value{
returns a formatted data frame to be combined with others to make the
  output of \code{hytek_swim_parse}
}
\description{
Sort data in lists of length 8 within \code{hytek_swim_parse}
}
