% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interleave_times.R
\name{interleave_times}
\alias{interleave_times}
\title{Helper for reading interleaving prelims and finals times}
\usage{
interleave_times(results, times, type = c("individual", "relay"))
}
\arguments{
\item{results}{a dataframe containing columns for minimum and maximum row number (usually `Row_Min` and `Row_Max`).  Times will be interleaved into this dataframe.}

\item{times}{a dataframe containing times in column 1 (or other values to be interleaved) and row numbers in column 2 (usually `Row_Numb`).}

\item{type}{either "individual" or "relay"}
}
\value{
a modified version of `results` with values from `times` interleaved on the basis of row number
}
\description{
Interleaves times based on row number ranges.
}
\seealso{
\code{interleave_times} is a helper function used in \code{\link{parse_hy3}}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
