% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/King200Breast.R
\docType{data}
\name{King200Breast}
\alias{King200Breast}
\title{Results for Lilly King's 200 Breaststrokes}
\format{
An object of class \code{"data.frame"}
}
\source{
\href{https://www.usaswimming.org/times/ncaa/ncaa-division-i}{NCAA Times Database}
}
\usage{
data(King200Breast)
}
\description{
Lilly King's 200 Breaststroke swims from her NCAA career
}
\keyword{datasets}
