% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lilly_King_2019_desc.R
\docType{data}
\name{Lilly_King_2019}
\alias{Lilly_King_2019}
\title{NCAA swims for Lilly King in 2018-2019 season}
\format{A data frame with 7 rows and 3 variables:
\describe{
  \item{Name}{Name of athlete, in this case Lilly King}
  \item{Time}{Time, in seconds}
  \item{Event}{name of the event}
}}
\usage{
Lilly_King_2019
}
\description{
A dataset of Indiana University breaststroke ace Lilly King's best times for each event she swam in the 2018-2019 NCAA season
}
\keyword{datasets}
