% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_breaker.R
\name{list_breaker}
\alias{list_breaker}
\title{Breaks out lists of lists by sub-list length}
\usage{
list_breaker(x, len)
}
\arguments{
\item{x}{a list of lists, with at least some sub-lists having length
\code{len}}

\item{len}{an numeric value for the length of sub-lists that
\code{list_breaker} should break out.  Must be a whole number.}
}
\value{
returns a list of lists, with all sub-lists having length \code{len}
}
\description{
XXXXXX
}
