% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_list_names.R
\name{list_to_list_names}
\alias{list_to_list_names}
\title{Initialize a named list of lists}
\usage{
list_to_list_names(list_of_names, value = 0)
}
\arguments{
\item{list_of_names}{a list of values, likely strings, to be the names of
sub-lists in a new list of lists}

\item{value}{a value to initialize the first element of all sub-lists to.
Defaults to \code{0}}
}
\value{
returns a list of lists with sub-list names from \code{list_of_names}
  and first elements from \code{value}.  Used inside \code{determine_entries}
}
\description{
Convert a single list to a list of lists, with the names of the lists taken
from the original list, \code{list_of_names}.  The new lists will all have a
single value, initialized as \code{value}.
}
