% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_transform.R
\name{list_transform}
\alias{list_transform}
\title{Transform list of lists into dataframe}
\usage{
list_transform(x)
}
\arguments{
\item{x}{a list of lists, with all sublists having the same length}
}
\value{
a dataframe where each sublist is a row and each element of that sublist is a column
}
\description{
Converts list of lists, with all sublists having the same number of elements into a dataframe where each sublist is a row and each element a column
}
\seealso{
\code{list_transform} is a helper function used inside of \code{swim_parse}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
