% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_pad.R
\name{na_pad}
\alias{na_pad}
\title{Pads shorter lists in a list-of-lists with \code{NA}s such that all lists are
the same length}
\usage{
na_pad(x, y)
}
\arguments{
\item{x}{a list of lists, with sub-lists having different lengths}

\item{y}{a list of the nubmer of \code{NA} values to append to each sub-list}
}
\value{
a list of lists with each sub-list the same length
}
\description{
Adds \code{NA} values to the end of each list in a list of lists such that
they all become the length of the longest list.  The longest list will not
have any \code{NA}s added to it.
}
