% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_times_parse.R
\name{reaction_times_parse}
\alias{reaction_times_parse}
\title{Pulls out reaction times from text}
\usage{
reaction_times_parse(text)
}
\arguments{
\item{text}{output from \code{read_results} followed by
\code{add_row_numbers}}
}
\value{
returns a data frame with reaction times and row numbers to
  eventually be recombined with swimming results inside \code{swim_parse}
}
\description{
Locates reaction times in text of results output from \code{read_results} and
their associated row numbers.  The resulting data frame is joined back into
results to include reaction times
}
\seealso{
\code{reaction_times_parse} is a helper function inside
  \code{\link{swim_parse}}
}
