% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Results.R
\name{read_htm}
\alias{read_htm}
\title{Read in html files of swimming results}
\usage{
read_htm(x, node_helper)
}
\arguments{
\item{x}{an .html, .htm or .aspx location containing swimming results.  Must be
formatted in a "normal" fashion - see vignette}

\item{node_helper}{receives \code{node} from \code{read_results}}
}
\value{
returns a list of results, with "read_results_flag" added as the
  first element of the list
}
\description{
Read in html files of swimming results
}
