% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Results.R
\name{read_hy3}
\alias{read_hy3}
\title{Read in hy3 files of swimming results}
\usage{
read_hy3(x)
}
\arguments{
\item{x}{an unzipped hy3 file containing swimming results.  Must be
formatted in a "normal" fashion - see vignette}
}
\value{
returns a list of results, with "read_results_flag" added as the
  first element of the list
}
\description{
Read in hy3 files of swimming results
}
