% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_pdf}
\alias{read_pdf}
\title{Read in pdf files of swimming results}
\usage{
read_pdf(x)
}
\arguments{
\item{x}{a .pdf or .aspx location containing swimming results.  Must be
formatted in a "normal" fashion - see vignette}
}
\value{
returns a list of results, with "read_results_flag" added as the
  first element of the list
}
\description{
Based on pdftools, this function can be temperamental
}
