% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_results_flag}
\alias{read_results_flag}
\title{used to indicate that results have been read in with \code{read_results}
prior to being parsed by \code{swim_parse}}
\usage{
read_results_flag(x)
}
\arguments{
\item{x}{a list of results, line by line}
}
\value{
returns list x, with "read_results_flag" added as the first element
  of the list
}
\description{
Used to insure that \code{read_results} has been run on a data source prior
to running \code{swim_parse}
}
