% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_entries.R
\name{replacement_entries}
\alias{replacement_entries}
\title{Replaces superseded rows}
\usage{
replacement_entries(x, index_number, row_to_add_replacement, e_df_replacement)
}
\arguments{
\item{x}{a data frame of entries, either df_helper_2 or Entries}

\item{index_number}{an integer denoting which element of e_df_replacement is
under test for removal.  Should be 1, 2, 3 or 4 depending on the minimum
number of entries}

\item{row_to_add_replacement}{a row containing an improved entry that should
be added to x}

\item{e_df_replacement}{a data frame of entries that may be replaced}
}
\value{
a data frame containing entries updated to include new rows from
  row_to_add_replacement and to not contain rows from e_df_replacement, based
  on index_number
}
\description{
Replaces superseded rows
}
