% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_format_helper.R
\name{sec_format_helper}
\alias{sec_format_helper}
\title{Helper function for formatting mm:ss.hh times as seconds}
\usage{
sec_format_helper(x)
}
\arguments{
\item{x}{A character vector of time(s) in swimming format (e.g. 1:35.93) to be converted to seconds (95.93)}
}
\description{
Helper function for formatting mm:ss.hh times as seconds
}
