% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splash_collect_splits.R
\name{splash_collect_splits}
\alias{splash_collect_splits}
\title{Collects Splash format splits}
\usage{
splash_collect_splits(df)
}
\arguments{
\item{df}{a data frame containing two columns, \code{V1} is row numbers and
\code{Dummy} as a string combining split distance and split time}
}
\value{
a data frame with three columns, \code{V1}, \code{Split_Distance} and
  \code{Split}
}
\description{
Collects splits and breaks them into a distance and a time, with a
corresponding row number
}
