% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splash_list_sorters.R
\name{splash_length_12_sort}
\alias{splash_length_12_sort}
\title{Sorts data in lists of length 12 within \code{splash_swim_parse}}
\usage{
splash_length_12_sort(x)
}
\arguments{
\item{x}{a list of lists with all sub-lists having length 12}
}
\value{
returns a formatted data frame to be combined with others to make the
  output of \code{splash_swim_parse}
}
\description{
XXXXXX
}
