% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sort.R
\name{split_sort}
\alias{split_sort}
\title{Sorts splits by performance row number}
\usage{
split_sort(x)
}
\arguments{
\item{x}{a list of splits}
}
\value{
a dataframe with \code{Row_Numb} as the first column.  Other columns are splits, by distance
}
\description{
Collects all splits associated with a paritcular performance (a swim) into a dataframe with the appropriate row number for that performance
}
\seealso{
\code{split_sort} is a helper function
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
