% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits_parse_ISL.R
\name{splits_parse_ISL}
\alias{splits_parse_ISL}
\title{Collects splits within \code{swim_parse_ISL}}
\usage{
splits_parse_ISL(text)
}
\arguments{
\item{text}{output of \code{read_results} with tow numbers appended by \code{add_row_numbers}}
}
\value{
returns a data frame with split times and row numbers
}
\description{
Takes the output of \code{read_results} and, inside of \code{swim_parse_ISL}, extracts split times and associated row numbers
}
\seealso{
\code{splits_parse_ISL} runs inside \code{\link{swim_parse_ISL}} on the output of \code{\link{read_results}} with row numbers from \code{\link{add_row_numbers}}
}
