% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splash_splits_parse.R
\name{splits_parse_splash_helper_1}
\alias{splits_parse_splash_helper_1}
\title{Produces data frames of splits within \code{swim_parse_splash} for Splash
results}
\usage{
splits_parse_splash_helper_1(data)
}
\arguments{
\item{data}{a list of lists containing splits and row numbers}
}
\value{
returns a data frame with split times and row numbers
}
\description{
Converts strings of splits and row numbers into data frames with a row number
column (V1) and a splits column (Split_XX)
}
\seealso{
\code{splits_parse_splash_helper_1} runs inside
  \code{\link{splits_parse_splash}}
}
