% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_splits_rename.R
\name{splits_rename_omega}
\alias{splits_rename_omega}
\title{Advances split names by one split_length}
\usage{
splits_rename_omega(x, split_len = split_length_omega)
}
\arguments{
\item{x}{a string to rename, from columns output by \code{splits_parse}}

\item{split_len}{distance for each split}
}
\value{
returns string iterated up by split_length
}
\description{
Used to adjust names of splits inside \code{swim_parse_omega} to account for
50 split not being correctly captured
}
\seealso{
\code{splits_rename_omega} runs inside \code{\link{swim_parse_omega}} on the
  output of \code{\link{splits_parse}}
}
