% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits_sort.R
\name{splits_sort}
\alias{splits_sort}
\title{Sorts splits by performance row number}
\usage{
splits_sort(x, min_row = minimum_row)
}
\arguments{
\item{x}{a list of splits}

\item{min_row}{the lowest row number}
}
\value{
a dataframe with \code{Row_Numb} as the first column.  Other columns are splits, by distance
}
\description{
Collects all splits associated with a particular performance (a swim) into a dataframe with the appropriate row number for that performance
}
\seealso{
\code{split_sort} is a helper function
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
