% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swim_Equation.R
\name{Swim_Equation}
\alias{Swim_Equation}
\title{Equation for computing swimming times from DPS, turn and start times}
\usage{
Swim_Equation(
  Swim_Time,
  Reaction_Time,
  Underwater_Time,
  Turn_Time,
  Cycle_Count,
  Stroke_Rate
)
}
\arguments{
\item{Swim_Time}{total time for a swim}

\item{Reaction_Time}{time leaving the blocks}

\item{Underwater_Time}{time spent underwater (dolphin kicking, pullouts, etc)}

\item{Turn_Time}{time spent completing turns}

\item{Cycle_Count}{total cycles}

\item{Stroke_Rate}{time per cycle}
}
\value{
whichever value is missing, solved for with respect to those provided
}
\description{
Takes a numeric item or list of numeric items representing seconds (eg 95.37) and converts to a character string or list of strings in swimming format ("1:35.37").
}
\examples{

}
\references{
from \url{https://www.usaswimming.org/docs/default-source/clinics/regional-coach-clinics/pittsburgh-rcc/race-stats---every-coach-can-use-everyday-for-every-athlete-rcc.pdf}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
