% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Swim_Parse.R
\name{Swim_Parse}
\alias{Swim_Parse}
\title{Formats swimming and diving data read in with Read_Results into dataframe}
\usage{
Swim_Parse(x, avoid = avoid_2, typo = typo_2, replacement = replacement_2)
}
\arguments{
\item{x}{the output or Read_Results}

\item{avoid}{a list of strings.  Rows in \code{x} containing these strings will not be included. For eample "Pool:", often used to label pool records, could be passed to \code{avoid}}

\item{typo}{a list of strings that are typos in the original results.  \code{Swim_Parse} is particularly sensitive to accidental double spaces, so "Central  High School", with two spaces between "Central" and "High" is a problem, which can be fixed.  Pass "Centeral High School" to \code{typo}.}

\item{replacement}{a list of fixes for the strings in \code{typo}.  Here one could pass "Central High School" (one space between "Central" and "High") to \code{replacement} fix the issue described in \code{typo}}
}
\value{
returns a dataframe with columns \code{Name}, \code{Place}, \code{Grade}, \code{School}, \code{Prelims_Time}, \code{Finals_Time}, \code{Points}, & \code{Event}.  Note all swims will have a \code{Finals_Time}, even if that time was actually swam in the prelims (i.e. a swimmer did not qualify for finals).  This is so that final results for an event can be generated from just one column.
}
\description{
Takes the output of Read_Results and cleans it.
}
\examples{
Swim_Parse(Read_Results("2008 NYSPHAA Federation Championship - 2_29_2008 to 3_1_2008.html", node = "pre"))
Swim_Parse(Read_Results("Texas-Florida-Indiana.pdf"))
}
\seealso{
\code{Swim_Parse} must be run on the output of \code{\link{Read_Results}}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
