% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwimmeR-defunct.R
\name{SwimmeR-defunct}
\alias{SwimmeR-defunct}
\title{Defunct functions in SwimmeR}
\description{
These functions have been made defunct (removed) from \code{SwimmeR}.
}
\details{
\itemize{ \item \code{\link{course_convert_DF}}: This function is defunct,
and has been removed from \code{SwimmeR}.  Instead
please use \code{course_convert(verbose = TRUE)} }
}
