% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwimmeR-deprecated.R
\name{SwimmeR-deprecated}
\alias{SwimmeR-deprecated}
\title{Deprecated functions in SwimmeR}
\description{
These functions still work but will be removed (defunct) in upcoming versions.
}
\details{
\itemize{ \item \code{\link{course_covert_df}}: This function is deprecated,
and will be removed in an updocming version this \code{SwimmeR}.  Instead
please use \code{course_convert(verbose = TRUE)} }
}
