% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwimR.R
\docType{package}
\name{SwimR}
\alias{SwimR}
\alias{SwimR-package}
\title{SwimR: A package for dealing with swimming times}
\description{
SwimR does two things.  First, it converts swimming times (performances) between the computationally useful 
    format of seconds, reported to the 100ths place (eg 95.37) and the conventional reporting format (1:35.37).
    Secondly SwimR converts between the various pool sizes used in competative swimming, namely 50m length (LCM), 25m length (SCM)
    and 25y length (SCY).
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
