% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_rescore.R
\name{tie_rescore}
\alias{tie_rescore}
\title{Rescore to account for ties}
\usage{
tie_rescore(df, point_values, lanes)
}
\arguments{
\item{df}{a dataframe with results from \code{swim_parse}, with places from \code{swim_place} and/or \code{dive_place}}

\item{point_values}{a named list of point values for each scoring place}

\item{lanes}{number of scoring lanes in the pool}
}
\value{
df modified so that places have been appended based on swimming time
}
\description{
Rescoring to average point values for ties.  Ties are placed as ties (both athletes get 2nd etc.)
}
\seealso{
\code{tie_rescore} is a helper function used inside of \code{results_score}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
