% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undo_interleave.R
\name{undo_interleave}
\alias{undo_interleave}
\title{Undoes interleaving of lists}
\usage{
undo_interleave(x)
}
\arguments{
\item{x}{a list to be un-interleaved}
}
\value{
a list comprising the interleaved components of \code{x} joined into
  one list
}
\description{
If two lists have been interleaved this function will return the lists
separated and then concatenated
}
\examples{
l <- c("A", "D", "B", "E", "C", "F")
undo_interleave(l)
}
