% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_entries.R
\name{update_rank_helper}
\alias{update_rank_helper}
\title{Create a one-line data frame containing an entry to be appended to an
in-progress data frame of all entries}
\usage{
update_rank_helper(
  rank_helper_2,
  e_rank_helper_2,
  k,
  e_helper,
  events_remaining_helper
)
}
\arguments{
\item{k}{an integer denoting which element of e_rank_helper is under
evaluation for addition.  Should be 1, 2, 3 or 4 depending on the minimum
number of entries}

\item{e_helper}{the event for which entries are being evaluated}

\item{events_remaining_helper}{a data frame with two columns, \code{Name} and
\code{Events_Remaining}}

\item{x}{a  master data frame of athlete ranks by event}

\item{e_rank_helper}{a data frame of candidate athlete entries to add to a
given event}
}
\value{
a one row data frame containing an improved entry
}
\description{
Create a one-line data frame containing an entry to be appended to an
in-progress data frame of all entries
}
