% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_age.R
\name{age_format_helper}
\alias{age_format_helper}
\title{Helper function for formatting yyy-mm ages as years, enables vectorization of
\code{age_format}}
\usage{
age_format_helper(x)
}
\arguments{
\item{x}{A character vector of age(s) in yyyy-mm format (e.g. 13-06) to be
converted to years (13.5)}
}
\description{
Helper function for formatting yyy-mm ages as years, enables vectorization of
\code{age_format}
}
