% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{Find the mode (most commonly occurring) element of a list}
\usage{
get_mode(x)
}
\arguments{
\item{x}{A list}
}
\value{
the element of \code{x} which appears most frequently.  Ties go to the lowest index, so the element which appears first.
}
\description{
Determines which element of list appears most frequently.  Based on \code{base::which.max}, so if multiple values appear with the same frequency will return the first one.  Ignores \code{NA} values.
In the context of swimming data is often used to clean team names, as in the Lilly King example below.
}
\examples{
a <- c("a", "a", "b", "c")
get_mode(a)
ab <- c("a", "a", "b", "b", "c") # returns "a", not "b", because although there are two of each "a" appears first
get_mode(ab)
a_na <- c("a", "a", NA, NA, "c")
get_mode(a_na)
numbs <- c(1, 1, 2, 2, 2, 3, NA)
get_mode(numbs)

Name <- c(rep("Lilly King", 5))
Team <- c(rep("IU", 2), "Indiana", "IUWSD", "Indiana University")
df <- data.frame(Name, Team, stringsAsFactors = FALSE)
df$Team <- get_mode(df$Team)

}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
