% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%notin\%}
\alias{\%notin\%}
\alias{\%!in\%}
\title{"Not in" function}
\usage{
x \%notin\% y

x \%!in\% y
}
\arguments{
\item{x}{a value}

\item{y}{a list of values}
}
\value{
a `TRUE` or `FALSE`
}
\description{
The opposite of `%in%`.  Returns `TRUE` if `x` is not `%in%` `y`.  Returns `FALSE` otherwise.
}
\examples{
"a" \%!in\% c("a", "b", "c")
"a" \%notin\% c("b", "c")

}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
