% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy3_times.R
\name{hy3_times}
\alias{hy3_times}
\title{Helper for reading prelims and finals times from Hy-Tek .hy3 files}
\usage{
hy3_times(file, type = c("prelims", "relay_prelims", "finals", "relay_finals"))
}
\arguments{
\item{file}{an output of read_results, from an .hy3 file}

\item{type}{type of times, either "prelims", "relay_prelims", "finals" or "relay_finals"}
}
\value{

}
\description{
XXX
}
\examples{
\dontrun{
 }
\dontrun{
 }
}
\seealso{
\code{parse_hy3} must be run on the output of \code{\link{read_results}}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
