% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lineup.R
\name{make_lineup_helper_2}
\alias{make_lineup_helper_2}
\title{Assign overpowered entries}
\usage{
make_lineup_helper_2(
  i,
  df_helper,
  in_progress_entries_df,
  events_competed_helper = Events_Competed,
  max_ind_entries = 3
)
}
\arguments{
\item{i}{a sequential list of numbers incremented by 1.  Used to index
function.}

\item{df_helper}{a data frame of all times to be entered for a given team.
Must contain column \code{Event} with the same event naming convention as
\code{op_df}, a column with name matching \code{time_col} containing times
or diving scores, and a column called \code{Name} containing athlete names}

\item{in_progress_entries_df}{a data frame containing the output of
\code{make_lineup_helper}, which is the minimum power set of entries}

\item{events_competed_helper}{a list of lists containing all the events a
given athlete is competing in.  Sub-lists are named with the athlete name.}

\item{max_ind_entries}{a numeric value denoting the maximum number of
individual events that may be entered by a single athlete}
}
\value{
a data frame containing entries updated to be as powerful as possible
}
\description{
Assign overpowered entries
}
