% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmss_format.R
\name{mmss_format}
\alias{mmss_format}
\title{Formatting seconds as mm:ss.tt}
\usage{
mmss_format(x)
}
\arguments{
\item{x}{A number of seconds to be converted to swimming format}
}
\description{
Takes a numeric item or list of numeric items representing seconds (eg 95.37) and converts to a character string or list of strings in swimming format ("1:35.37").
}
\seealso{
\code{\link{sec_format}} \code{mmss_format} is the reverse of \code{sec_format}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
