% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_entries.R
\name{replacement_entries}
\alias{replacement_entries}
\title{Replaces superseded rows}
\usage{
replacement_entries(x, index_number, row_to_add_replacement, e_df_replacement)
}
\arguments{
\item{i}{a sequential list of numbers incremented by 1.  Used to index
function.}

\item{df_helper}{a data frame of times for the team to be entered.  Must
contain column \code{Event} with the same event naming convention as
\code{op_df}, a column with name matching \code{time_col} containing times
or diving scores, and a column called \code{Name} containing athlete names}

\item{op_df_helper}{a data frame containing the opponent lineup.  Must
contain column \code{Event} with the same event naming convention as
\code{df}, a column with name matching \code{time_col} containing times or
diving scores, and a column called \code{Name} containing athlete names}

\item{max_ind_entries}{a numeric value denoting the maximum number of
individual events that may be entered by a single athlete}
}
\value{
xxx
}
\description{
Replaces superseded rows
}
