% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_score.R
\name{results_score}
\alias{results_score}
\title{Scores a swim meet}
\usage{
results_score(
  results,
  events,
  meet_type = c("timed_finals", "prelims_finals"),
  lanes = c(4, 6, 8, 10),
  scoring_heats = c(1, 2, 3),
  point_values
)
}
\arguments{
\item{results}{an output from \code{swim_parse}}

\item{events}{list of events}

\item{meet_type}{how to score based on \code{timed_finals}, where any place is possible, or \code{prelims_finals} where athletes are locked into heats for scoring purposes}

\item{lanes}{number of lanes in to the pool, for purposes of heat}

\item{scoring_heats}{number of heats which score (if 1 only A final scores, if 2 A and B final score etc.)}

\item{point_values}{a list of point values for each scoring place}
}
\value{
results with point values in a column called \code{Points}
}
\description{
Used to add a \code{Points} column with point values for each place.  Can either score "timed finals" type meets where any athlete can get any place, or "prelims-finals", type meets, where placing is restricted by prelim performance.
}
\examples{
\dontrun{read_results("http://www.nyhsswim.com/Results/Boys/2008/NYS/Single.htm", node = "pre")}

}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
