% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits_reform.R
\name{splits_reform}
\alias{splits_reform}
\title{Adds together splits and compares to listed finals time to see if they match.}
\usage{
splits_reform(df)
}
\arguments{
\item{df}{a dataframe output from \code{swim_parse} creates with \code{splits = TRUE}}
}
\value{
a dataframe with a column \code{not_matching} containing \code{TRUE} if the splits for that swim match the finals time and \code{FALSE} if they do not
}
\description{
Used in testing the workings for \code{split_parse} inside test-splits.R.
Note that even properly handled splits may not match the finals time due to issues in the source material.
Sometimes splits aren't fully recorded in the source.
Some relays also will not match due to the convention of reporting splits by swimmer (see vignette for more details).
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
