% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swim_place.R
\name{swim_place}
\alias{swim_place}
\title{Adds places to swimming results}
\usage{
swim_place(df, max_place)
}
\arguments{
\item{df}{a dataframe with results from \code{swim_parse}, including only swimming results (not diving)}

\item{max_place}{highest place value that scores}
}
\value{
df modified so that places have been appended based on swimming time
}
\description{
Places are awarded on the basis of time, with fastest (lowest) time winning.  Ties are placed as ties (both athletes get 2nd etc.)
}
\seealso{
\code{swim_place} is a helper function used inside of \code{results_score}
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
