% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwimR.R
\docType{package}
\name{SwimmeR}
\alias{SwimmeR}
\alias{SwimmeR-package}
\title{SwimmeR: A package for working with swimming times}
\description{
The goal for of the 'SwimmeR' package is to provide means of acquirring, and then analyizing data from swimming (and diving) competitions.  To that end 'SwimmeR' allows results to be read in from .html sources, like Hy-Tek real time results pages, .pdf files, and now (preliminarily) .hy3 files.  Once read in, 'SwimmeR' can convert swimming times (performances) between the computationally useful format of seconds reported to the 100ths place (e.g. 95.37), and the conventional reporting format (1:35.37) used in the swimming community.  'SwimmeR' can also score meets in a variety of formats with user defined point values, convert times between courses (LCM, SCM, SCY) and draw single elimination brackets, as well as providing a suite of tools for working cleaning swimming data.  This is a developmental package, not yet mature.
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
