% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwimR.R
\docType{package}
\name{SwimmeR}
\alias{SwimmeR}
\alias{SwimmeR-package}
\title{SwimmeR: A package for dealing with swimming times}
\description{
There are two goals for 'SwimmeR' as presently constructed.  First, it converts swimming times (performances) between the computationally useful 
    format of seconds, reported to the 100ths place (eg 95.37) and the conventional reporting format (1:35.37) used in the swimming community.
    Secondly 'SwimmeR' converts times between the various pool sizes used in competitive swimming, namely 50m length (LCM), 25m length (SCM)
    and 25y length (SCY).
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
