% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwimR.R
\docType{package}
\name{SwimR}
\alias{SwimR}
\alias{SwimR-package}
\title{SwimR: A package for dealing with swimming times}
\description{
What the package does (one paragraph).
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
